Fix RowHit/RowOpen lambdas in LPDDR5
The use of the RowOpen/RowHit lambdas to detect the
bank state during a RD/WR operation does not consider
the Pre-Opened state, specific to LPDDR5. This results
in assuming the default case, which provokes an execution
error.
This commit fixes this behavior by using custom lambdas
for LPDDR5, to include the additional state.

A similar fix was provided by the PSAL-POSTECH group in
https://github.com/PSAL-POSTECH/ramulator2, commit
bb1d54f5117a6232f13724bb40f903776645a737.

Fixes issue #66.